swagger-config: настройка swagger,который не запускался
