+++ src/main/java/ru/dynamica/library/LibraryApplication.java
@@ -9,3 +9,4 @@ public class LibraryApplication {
         SpringApplication.run(LibraryApplication.class, args);
     }
 }
+
