+++ src/main/java/ru/dynamica/library/config/WebMvcConfiguration.java
@@ -1,0 +1,17 @@
+package ru.dynamica.library.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+/**
+ * Класс с конфигурированием {@code Spring Web}.
+ */
+@Configuration
+public class WebMvcConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**").allowedMethods("*");
+    }
+}
