+++ src/main/java/ru/dynamica/library/model/dto/AuthorDto.java
@@ -1,6 +1,5 @@
-package ru.dynamica.library.model.dto.common;
+package ru.dynamica.library.model.dto;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,15 +13,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class AuthorDto {
-    @Schema(description = "Id автора", example = "1")
     private Long id;
 
-    @Schema(description = "Имя", example = "Фёдор")
     private String name;
 
-    @Schema(description = "Фамилия", example = "Достоевский")
     private String surname;
 
-    @Schema(description = "Отчество", example = "Михайлович")
     private String fathername;
 }
