+++ src/main/java/ru/dynamica/library/model/dto/BookDto.java
@@ -1,6 +1,5 @@
-package ru.dynamica.library.model.dto.common;
+package ru.dynamica.library.model.dto;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,15 +13,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class BookDto {
-    @Schema(description = "Id книги", example = "1")
     private Long id;
 
-    @Schema(description = "ISNB книги", example = "978-2-266-11156-0")
     private String ISBN;
 
-    @Schema(description = "Название книги", example = "Братья Карамазовы")
     private String title;
 
-    @Schema(description = "Автор книги", example = "1")
     private Long authorId;
 }
