--- src/main/java/ru/dynamica/library/model/dto/response/ApiResponseDto.java
@@ -1,24 +1,0 @@
-package ru.dynamica.library.model.dto.response;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * Ответ API.
- *
- * @param <T> тип DTO.
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ApiResponseDto<T> {
-
-    private ApiError error;
-
-    @Schema(description = "Информация ответа.")
-    private T content;
-}
