+++ src/main/java/ru/dynamica/library/service/impl/BookRentServiceImpl.java
@@ -1,15 +1,16 @@
 package ru.dynamica.library.service.impl;
 
 import org.springframework.stereotype.Service;
-import ru.dynamica.library.model.dto.common.BookDto;
+import ru.dynamica.library.model.dto.BookDto;
 import ru.dynamica.library.service.BookRentService;
 
+import java.util.Collections;
 import java.util.List;
 
 @Service
 public class BookRentServiceImpl implements BookRentService {
     @Override
     public List<BookDto> getBooks() {
-        return null;
+        return Collections.emptyList();
     }
 }
